requirements.txt
requests>=2.28.0
discord.py>=2.3.2
google-generativeai>=0.8.0
python-dotenv>=1.0.0